+++ zcash_transparent/src/builder.rs
@@ -99,8 +99,7 @@ impl fmt::Display for Error {
     }
 }
 
-#[cfg(feature = "std")]
-impl std::error::Error for Error {}
+impl core::error::Error for Error {}
 
 /// A set of transparent signing keys.
 ///