+++ zcash_transparent/src/coinbase.rs
@@ -60,5 +60,4 @@ impl fmt::Display for Error {
     }
 }
 
-#[cfg(feature = "std")]
-impl std::error::Error for Error {}
+impl core::error::Error for Error {}